Random Word Gen V.1
Added a new project called "Random Word Gen". It is a simple project to get random words from word list "words.txt".

Changed "int" to "float" in Temp Convert so it can convert decimal number.

Also fixed some previous code.

+++ Class/Confirmed/5 Celcius to Farenheit.py
@@ -1,4 +1,4 @@
-C = int(input("Enter the value of Celsius: "))
+C = float(input("Enter the value of Celsius: "))
 
 F = (C*9/5)+32
